Crate preempt_rwlock

Structs

OwnedPreemptibleRwLockReadGuard

PreemptibleRwLock

A RW-lock which also supports a way for pending writers to request that readers get preempted, via preempted() method on the read guard that readers can tokio::select! on.

PreemptibleRwLockReadGuard

PreemptibleRwLockWriteGuard